REFACTOR: Add `WebViewViewControllerProtocol`, `WebViewPresenterProtocol` and `WebViewPresenter`; Transfer URLRequest construction to `WebViewPresenter`
