+++ ImageFeed/Authentication/WebViewPresenter.swift
@@ -1,0 +1,28 @@
+import UIKit
+
+final class WebViewPresenter: WebViewPresenterProtocol {
+    weak var view: WebViewViewControllerProtocol?
+    
+    func viewDidLoad() {
+        guard var urlComponents = URLComponents(string: WebViewConstants.unsplashAuthorizeURLString) else {
+            print("Error: Unable to create URLComponents from unsplashAuthorizeURLString")
+            return
+        }
+        
+        urlComponents.queryItems = [
+            URLQueryItem(name: "client_id", value: Constants.accessKey),
+            URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
+            URLQueryItem(name: "response_type", value: "code"),
+            URLQueryItem(name: "scope", value: Constants.accessScope)
+        ]
+        
+        guard let url = urlComponents.url else {
+            print("Error: Unable to create URL from urlComponents")
+            return
+        }
+        
+        let request = URLRequest(url: url)
+        
+        view?.load(request: request)
+    }
+}
