+++ ImageFeed/Protocols/WebViewPresenterProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol WebViewPresenterProtocol {
+    var view: WebViewViewControllerProtocol? { get set }
+    func viewDidLoad()
+}
