+++ ImageFeed/Protocols/WebViewViewControllerProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol WebViewViewControllerProtocol: AnyObject {
+    var presenter: WebViewPresenterProtocol? { get set }
+    func load(request: URLRequest)
+}
